Website_apresentacao/backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-CORS==3.0.10
python-dotenv==1.0.0
spacy==3.5.3
psycopg2-binary==2.9.6
Werkzeug==2.2.3
numpy==1.23.5